Minor improvements to the traditional Synergy windows build script.

+++ TraditionalSynergy/Windows/clean.bat
@@ -49,19 +49,21 @@ echo =================================
 set ROOT=%~dp0
 pushd "%ROOT%"
 
-del /q hdr\*.*
-del /q obj\*.*
-del /q exe\*.dbr
-del /q exe\*.elb
-del /q exe\*.chm
-del /q exe\codegen.bat
-del /q exe\codegend.bat
-del /q exe\mapprep.bat
-del /q exe\rpsinfo.bat
-del /q exe\setenv.bat
+if exist hdr\*.* del /q hdr\*.*
+if exist obj\*.* del /q obj\*.*
+if exist exe\*.dbr del /q exe\*.dbr
+if exist exe\*.elb del /q exe\*.elb
+if exist exe\*.chm del /q exe\*.chm
+if exist exe\codegen.bat del /q exe\codegen.bat
+if exist exe\codegend.bat del /q exe\codegend.bat
+if exist exe\mapprep.bat del /q exe\mapprep.bat
+if exist exe\rpsinfo.bat del /q exe\rpsinfo.bat
+if exist exe\createfile.bat del /q exe\createfile.bat
+if exist exe\setenv.bat del /q exe\setenv.bat
+if exist exe\DefaultButtons.xml del /q exe\DefaultButtons.xml
 
-rmdir /q hdr
-rmdir /q obj
+if exist hdr\. rmdir /q hdr
+if exist obj\. rmdir /q obj
 
 popd
 endlocal
